TDS 01 Basic movement!
Created basic Click-To-Move system for the player. You can also hold it down if you want to move non-stop. Thus far works great, no clue on the impact in terms of performance.
Since it uses the navmesh system, it also has no issues going up terrain or around corners. Speed can obviously be adjusted if needed.
Also created a very basic AI that moves towards the player with the same navmesh system as long as the player is within a X radius of the enemy/monster.