TDS 02 Basic movement animations!
Today, I've added two very, very basic animations to the soldier!
Animation one is a simple idle animation and the second one is a running animation. It looks perfect for when the soldier will have a gun or anything else in his hands. I'm currently not quite sure how I'll set that up, I think I'll have to mess around with sockets and what not, but currently it's not that bad.
One bug that I have found is with my current rotation system. The character rotates to the vector direction of where the user clicked (aka the target location), so if the character must move around objects using his navigation system, he is already facing the target without actually seeing it, which makes it look kind of weird. I might look into it in the future, but for now it's good.