TDS 03 Enemy animations.
The FirstEnemy character has received some animations as well as a completely new mesh. They're still placeholder, but I wanted to do it in order to somewhat understand how the animations work and reinforce my already existing knowledge.
The monster received a idle and running animation just like the soldier before, but I did as well add a attacking animation, but it's not used anywhere.
This monster character had a way easier time with it's skeleton unlike the soldier.
Maybe in the future I could make my own models and skeletons, but that's just a dream for now.