TDS 06
-Removed the sword blueprint and its assets.
-Created a fully multiplayer ready movement system.

The movement system works as such:
There is a Player Proxy pawn that has its own controller. This Player Proxy then spawns the actual player actor (the one with the model, weapons etc.) and it's controller.
Proxy Controller gets click vector and then passes it on to the real players controller which then just does a simple move to location through the NavMesh in the level.