TDS 08 Enemy attack, animation and weapon pack!
Created a way for the monster to attack. It uses a straight bar which is a box collision. Using some logic and a "event any damage" node, managed to make the attack system fully change the players health. On top of that it also works flawlessly on multiplayer.
Managed to get an attack animation added to the monster.
**note** Should probably change attack speed to match animation.

Added a UE4 weapons pack from the marketplace in hopes to start working on the weapons soon.