TDS 09 Deleted useless files, spawning weapon, shooting weapon.
Deleted two useless files, those being AIController_C and PlayerController_PC. These are useless since the actual player uses completely different Controllers now.

Managed to get a weapon of Wep_BaseClass to spawn in the hands of the player.
The Uzi weapon now has a Shooting Socket added to it. This is the socket where projectiles will spawn when the player shoots.
Managed to get the gun to actually fire from the socket as well, which is very nice.
Had to screw around quite a bit with various ways of casting things between the player, weapon and the real playercontroller (the one for the proxy, which actually takes input). Gladly, it all worked out.

-- Due to the new projectiles being spawned, the First_Enemy no longer works due to issues with the additional actors in its aggro sphere!

Next up: Fixing the shooting, so the bullets fly the appropriate way.